+++ LaHuerta/backend/tipo_contenedor/views.py
@@ -4,6 +4,7 @@ from rest_framework import status
 from .serializers import ContainerTypeSerializer
 from .repositories import ContainerTypeRepository
 from .exceptions import ContainerHasProductsException, ContainerNotFoundException
+from producto.repositories import ProductRepository
 class ContainerTypeViewSet(viewsets.ModelViewSet):
     '''
     Gestión de Tipos de Contenedores
@@ -61,9 +62,23 @@ class ContainerTypeViewSet(viewsets.ModelViewSet):
         Elimina un tipo de contenedor.
         '''
         try:
+            container_type = self.container_type_repository.get_container_by_id(pk)
+
+            if not container_type:
+                raise ContainerNotFoundException('El tipo de contenedor seleccionado no existe')
+
+            product_repository = ProductRepository()
+            related_products = product_repository.verify_products_with_container_type_id(pk)
+
+            if related_products:
+                raise ContainerHasProductsException('El tipo de contenedor seleccionado tiene productos asociados')
+            
             self.container_type_repository.destroy_container_type(pk)
             return Response({'message': 'Tipo de contenedor eliminado exitosamente'}, status=status.HTTP_204_NO_CONTENT)
+        
         except ContainerHasProductsException as e:
             return Response({'error': str(e)}, status=status.HTTP_400_BAD_REQUEST)
         except ContainerNotFoundException as e:
             return Response({'error': str(e)}, status=status.HTTP_404_NOT_FOUND)
+        except Exception as e:
+            return Response({'error': 'Ocurrió un error inesperado'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
